I have now addressed all seven issues
