+++ src/ru/astramanga/src/eu/kanade/tachiyomi/extension/ru/astramanga/AstraManga.kt
@@ -10,14 +10,10 @@ import eu.kanade.tachiyomi.source.online.HttpSource
 import keiyoushi.annotation.Source
 import keiyoushi.network.rateLimit
 import keiyoushi.utils.parseAs
-import keiyoushi.utils.tryParse
 import okhttp3.HttpUrl.Companion.toHttpUrl
 import okhttp3.Request
 import okhttp3.Response
 import rx.Observable
-import java.text.SimpleDateFormat
-import java.util.Locale
-import java.util.TimeZone
 
 @Source
 abstract class AstraManga : HttpSource() {
@@ -65,9 +61,7 @@ abstract class AstraManga : HttpSource() {
                     ?.let { builder.addQueryParameter("type", it) }
                 is StatusFilter -> filter.selected.takeIf { it.isNotEmpty() }
                     ?.let { builder.addQueryParameter("status", it) }
-                is SortFilter -> if (query.isBlank()) {
-                    builder.addQueryParameter("sort", filter.selected)
-                }
+                is SortFilter -> builder.addQueryParameter("sort", filter.selected)
                 is GenreFilter -> filter.state.filter { it.state }
                     .forEach { builder.addQueryParameter("genres", it.id) }
                 else -> {}
@@ -80,22 +74,22 @@ abstract class AstraManga : HttpSource() {
 
     private fun searchParse(response: Response): MangasPage {
         val data = response.parseAs<SearchResponse>().data
-        val mangas = data.titles.map { it.toSManga() }
+        val mangas = data.titles.map { it.toSManga(mediaUrl) }
         return MangasPage(mangas, data.currentPage < data.totalPages)
     }
 
     // ============================== Details ===============================
 
-    override fun mangaDetailsRequest(manga: SManga): Request = GET("$apiUrl/titles/${manga.slug()}", headers)
+    override fun mangaDetailsRequest(manga: SManga): Request = GET("$apiUrl/titles/${manga.url}", headers)
 
-    override fun getMangaUrl(manga: SManga): String = baseUrl + manga.url
+    override fun getMangaUrl(manga: SManga): String = "$baseUrl/manga/${manga.url}"
 
-    override fun mangaDetailsParse(response: Response): SManga = response.parseAs<TitleDetailResponse>().data.toSMangaDetails()
+    override fun mangaDetailsParse(response: Response): SManga = response.parseAs<TitleDetailResponse>().data.toSMangaDetails(mediaUrl)
 
     // ============================== Chapters ==============================
 
     override fun fetchChapterList(manga: SManga): Observable<List<SChapter>> = Observable.fromCallable {
-        val slug = manga.slug()
+        val slug = manga.url
 
         // Resolve the numeric title id (the titles endpoint accepts the slug).
         val titleId = client.newCall(GET("$apiUrl/titles/$slug", headers)).execute()
@@ -118,13 +112,16 @@ abstract class AstraManga : HttpSource() {
             .sortedByDescending { it.chapter_number }
     }
 
-    override fun getChapterUrl(chapter: SChapter): String = baseUrl + chapter.url
+    override fun getChapterUrl(chapter: SChapter): String {
+        val (slug, number, id) = chapter.url.split("/", limit = 3)
+        return "$baseUrl/manga/$slug/read/$number?chapterId=$id"
+    }
 
     override fun chapterListParse(response: Response): List<SChapter> = throw UnsupportedOperationException()
 
     // =============================== Pages ================================
 
-    override fun pageListRequest(chapter: SChapter): Request = GET("$apiUrl/chapters/${chapter.chapterId()}/pages", headers)
+    override fun pageListRequest(chapter: SChapter): Request = GET("$apiUrl/chapters/${chapter.url.substringAfterLast('/')}/pages", headers)
 
     override fun pageListParse(response: Response): List<Page> {
         val pages = response.parseAs<PagesResponse>().data.pages
@@ -143,79 +140,8 @@ abstract class AstraManga : HttpSource() {
         GenreFilter(),
     )
 
-    // ============================== Mappers ===============================
-
-    private fun SManga.slug(): String = url.substringAfter("/manga/").substringBefore("?").substringBefore("#")
-    private fun SChapter.chapterId(): String = url.substringAfter("chapterId=").substringBefore("&").substringBefore("#")
-
-    private fun TitleDto.coverUrl(): String? {
-        val path = coverImage ?: coverVersions?.mid ?: coverVersions?.high ?: return null
-        return "$mediaUrl/$path"
-    }
-
-    private fun TitleDto.toSManga(): SManga = SManga.create().apply {
-        url = "/manga/$slug"
-        title = name
-        thumbnail_url = coverUrl()
-    }
-
-    private fun TitleDto.toSMangaDetails(): SManga = SManga.create().apply {
-        url = "/manga/$slug"
-        title = name
-        thumbnail_url = coverUrl()
-        author = publishingHouse?.name ?: publishers?.firstOrNull()?.name
-        description = buildString {
-            secondaryName?.takeIf { it.isNotBlank() }?.let { appendLine("Альт. название: $it") }
-            alternativeNames?.takeIf { it.isNotEmpty() }
-                ?.let { appendLine("Другие названия: ${it.joinToString()}") }
-            year?.let { appendLine("Год выпуска: $it") }
-            if (isNotEmpty()) appendLine()
-            append(this@toSMangaDetails.description?.trim().orEmpty())
-        }.trim()
-        val tagNames = buildList {
-            this@toSMangaDetails.type?.let { add(typeName(it)) }
-            genres?.forEach { g -> g.name?.let { add(it) } }
-            tags?.forEach { t -> t.name?.let { add(it) } }
-        }
-        genre = tagNames.filter { it.isNotBlank() }.distinct().joinToString()
-        status = parseStatus(this@toSMangaDetails.status)
-    }
-
-    private fun ChapterDto.numberStr(): String = number.toString().removeSuffix(".0")
-
-    private fun ChapterDto.toSChapter(slug: String): SChapter = SChapter.create().apply {
-        url = "/manga/$slug/read/${numberStr()}?chapterId=$id"
-        name = buildString {
-            if (volumeNumber != null) append("Том $volumeNumber ")
-            append("Глава ${numberStr()}")
-            this@toSChapter.name?.takeIf { it.isNotBlank() }?.let { append(" — $it") }
-        }
-        chapter_number = number
-        date_upload = DATE_FORMAT.tryParse(publishedAt)
-    }
-
-    private fun parseStatus(status: String?): Int = when (status) {
-        "ongoing" -> SManga.ONGOING
-        "completed" -> SManga.COMPLETED
-        "paused" -> SManga.ON_HIATUS
-        "frozen", "discontinued" -> SManga.CANCELLED
-        else -> SManga.UNKNOWN
-    }
-
-    private fun typeName(type: String): String = when (type) {
-        "manga" -> "Манга"
-        "manhwa" -> "Манхва"
-        "manhua" -> "Маньхуа"
-        "comics" -> "Комикс"
-        else -> type
-    }
-
     companion object {
         private const val PAGE_SIZE = 30
         private const val CHAPTERS_PAGE_SIZE = 10000
-
-        private val DATE_FORMAT = SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss", Locale.ROOT).apply {
-            timeZone = TimeZone.getTimeZone("UTC")
-        }
     }
 }
